Move getParticipationWithResults to ProgrammingExerciseParticipationService
