+++ src/main/java/de/tum/cit/aet/artemis/programming/service/ProgrammingExerciseParticipationService.java
@@ -8,6 +8,7 @@ import java.io.IOException;
 import java.util.List;
 import java.util.Optional;
 
+import jakarta.annotation.Nullable;
 import jakarta.validation.constraints.NotNull;
 
 import org.apache.commons.io.FileUtils;
@@ -62,9 +63,18 @@ public class ProgrammingExerciseParticipationService {
 
     private final GitService gitService;
 
+    private final TemplateProgrammingExerciseParticipationRepository templateProgrammingExerciseParticipationRepository;
+
+    private final SolutionProgrammingExerciseParticipationRepository solutionProgrammingExerciseParticipationRepository;
+
+    private final ProgrammingExerciseStudentParticipationRepository programmingExerciseStudentParticipationRepository;
+
     public ProgrammingExerciseParticipationService(SolutionProgrammingExerciseParticipationRepository solutionParticipationRepository,
             TemplateProgrammingExerciseParticipationRepository templateParticipationRepository, ProgrammingExerciseStudentParticipationRepository studentParticipationRepository,
-            ParticipationRepository participationRepository, TeamRepository teamRepository, GitService gitService, Optional<VersionControlService> versionControlService) {
+            ParticipationRepository participationRepository, TeamRepository teamRepository, GitService gitService, Optional<VersionControlService> versionControlService,
+            TemplateProgrammingExerciseParticipationRepository templateProgrammingExerciseParticipationRepository,
+            SolutionProgrammingExerciseParticipationRepository solutionProgrammingExerciseParticipationRepository,
+            ProgrammingExerciseStudentParticipationRepository programmingExerciseStudentParticipationRepository) {
         this.studentParticipationRepository = studentParticipationRepository;
         this.solutionParticipationRepository = solutionParticipationRepository;
         this.templateParticipationRepository = templateParticipationRepository;
@@ -72,6 +82,9 @@ public class ProgrammingExerciseParticipationService {
         this.teamRepository = teamRepository;
         this.versionControlService = versionControlService;
         this.gitService = gitService;
+        this.templateProgrammingExerciseParticipationRepository = templateProgrammingExerciseParticipationRepository;
+        this.solutionProgrammingExerciseParticipationRepository = solutionProgrammingExerciseParticipationRepository;
+        this.programmingExerciseStudentParticipationRepository = programmingExerciseStudentParticipationRepository;
     }
 
     /**
@@ -350,5 +363,37 @@ public class ProgrammingExerciseParticipationService {
             log.error("Could not get commit infos for test repository with participation id {}", participation.getId());
             return List.of();
         }
+    }
+
+    /**
+     * Returns the matching template, solution or student participation for a given build plan key.
+     *
+     * @param planKey the build plan key
+     * @return the matching participation
+     */
+    @Nullable
+    public ProgrammingExerciseParticipation getParticipationWithResults(String planKey) {
+        // we have to support template, solution and student build plans here
+        if (planKey.endsWith("-" + BuildPlanType.TEMPLATE.getName())) {
+            return templateProgrammingExerciseParticipationRepository.findByBuildPlanIdWithResults(planKey).orElse(null);
+        }
+        else if (planKey.endsWith("-" + BuildPlanType.SOLUTION.getName())) {
+            return solutionProgrammingExerciseParticipationRepository.findByBuildPlanIdWithResults(planKey).orElse(null);
+        }
+        List<ProgrammingExerciseStudentParticipation> participations = programmingExerciseStudentParticipationRepository
+                .findWithResultsAndExerciseAndTeamStudentsByBuildPlanId(planKey);
+        ProgrammingExerciseStudentParticipation participation = null;
+        if (!participations.isEmpty()) {
+            participation = participations.getFirst();
+            if (participations.size() > 1) {
+                // in the rare case of multiple participations, take the latest one.
+                for (ProgrammingExerciseStudentParticipation otherParticipation : participations) {
+                    if (otherParticipation.getInitializationDate().isAfter(participation.getInitializationDate())) {
+                        participation = otherParticipation;
+                    }
+                }
+            }
+        }
+        return participation;
     }
 }
